Move series sharding code to util package.
